Move classes to `operations` package
