+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/DefaultPlannedTransformStep.java
@@ -18,6 +18,7 @@ package org.gradle.api.internal.artifacts.transform;
 
 import org.gradle.internal.taskgraph.CalculateTaskGraphBuildOperationType;
 import org.gradle.internal.taskgraph.NodeIdentity;
+import org.gradle.operations.dependencies.transforms.PlannedTransformStepIdentity;
 
 import java.util.List;
